pyproject.toml
[project]
name = "be"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "bcrypt>=5.0.0",
    "cryptography>=46.0.3",
    "flask>=3.1.2",
    "flask-bcrypt>=1.0.1",
    "flask-cors>=6.0.1",
    "flask-jwt-extended>=4.7.1",
    "flask-sqlalchemy>=3.1.1",
    "psycopg2-binary>=2.9.11",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.2.1",
]
